+++ framework/Source/CPTImage.m
@@ -481,7 +481,7 @@ CPTImageSlices;
            CGImageGetBytesPerRow(selfCGImage) +
            CGImageGetBitmapInfo(selfCGImage) +
            CGImageGetShouldInterpolate(selfCGImage) +
-           (NSUInteger)(CGImageGetRenderingIntent(selfCGImage) * self.scale);
+           (NSUInteger)(CGImageGetRenderingIntent(selfCGImage)) * (NSUInteger)self.scale;
 }
 
 /// @endcond
